tec: Extract migrations actions in a dedicated controller
I want to add more commands related to migrations so it makes sense to
extract these ones in a different controller.
