+++ src/cli/System.php
@@ -114,7 +114,8 @@ class System
         $migrations_version_path = $data_path . '/migrations_version.txt';
 
         if (file_exists($migrations_version_path)) {
-            $response = $this->migrate();
+            $migrations_controller = new Migrations();
+            $response = $migrations_controller->apply();
         } else {
             $response = $this->init();
         }
@@ -162,61 +163,6 @@ class System
     }
 
     /**
-     * Execute the migrations under src/migrations/. The version is stored in
-     * the data/migrations_version.txt file.
-     *
-     * @return \Minz\Response
-     */
-    private function migrate()
-    {
-        $app_path = \Minz\Configuration::$app_path;
-        $data_path = \Minz\Configuration::$data_path;
-        $migrations_path = $app_path . '/src/migrations';
-        $migrations_version_path = $data_path . '/migrations_version.txt';
-
-        $migration_version = @file_get_contents($migrations_version_path);
-        if ($migration_version === false) {
-            return Response::text(500, 'Cannot read the migrations version file.'); // @codeCoverageIgnore
-        }
-
-        $migrator = new \Minz\Migrator($migrations_path);
-        $migration_version = trim($migration_version);
-        if ($migration_version) {
-            $migrator->setVersion($migration_version);
-        }
-
-        if ($migrator->upToDate()) {
-            return Response::text(200, 'Your system is already up to date.');
-        }
-
-        $results = $migrator->migrate();
-
-        $new_version = $migrator->version();
-        $saved = @file_put_contents($migrations_version_path, $new_version);
-        if ($saved === false) {
-            $text = "Cannot save the migrations version file (version: {$version})."; // @codeCoverageIgnore
-            return Response::text(500, $text); // @codeCoverageIgnore
-        }
-
-        $has_error = false;
-        $results_as_text = [];
-        foreach ($results as $migration => $result) {
-            if ($result === false) {
-                $result = 'KO';
-            } elseif ($result === true) {
-                $result = 'OK';
-            }
-
-            if ($result !== 'OK') {
-                $has_error = true;
-            }
-
-            $results_as_text[] = "{$migration}: {$result}";
-        }
-        return Response::text($has_error ? 500 : 200, implode("\n", $results_as_text));
-    }
-
-    /**
      * Execute seeds code in src/seeds.php.
      *
      * @response 201 If seeds file don’t exist
@@ -237,60 +183,5 @@ class System
             $error = (string)$e;
             return Response::text(500, "Seeds can’t be loaded:\n{$error}");
         }
-    }
-
-    /**
-     * Execute the rollback of the latest migrations.
-     *
-     * @request_param integer steps (default is 1)
-     *
-     * @param \Minz\Request $request
-     *
-     * @return \Minz\Response
-     */
-    public function rollback($request)
-    {
-        $app_path = \Minz\Configuration::$app_path;
-        $data_path = \Minz\Configuration::$data_path;
-        $migrations_path = $app_path . '/src/migrations';
-        $migrations_version_path = $data_path . '/migrations_version.txt';
-
-        $migration_version = @file_get_contents($migrations_version_path);
-        if ($migration_version === false) {
-            return Response::text(500, 'Cannot read the migrations version file.'); // @codeCoverageIgnore
-        }
-
-        $migrator = new \Minz\Migrator($migrations_path);
-        $migration_version = trim($migration_version);
-        if ($migration_version) {
-            $migrator->setVersion($migration_version);
-        }
-
-        $steps = $request->paramInteger('steps', 1);
-        $results = $migrator->rollback($steps);
-
-        $new_version = $migrator->version();
-        $saved = @file_put_contents($migrations_version_path, $new_version);
-        if ($saved === false) {
-            $text = "Cannot save the migrations version file (version: {$version})."; // @codeCoverageIgnore
-            return Response::text(500, $text); // @codeCoverageIgnore
-        }
-
-        $has_error = false;
-        $results_as_text = [];
-        foreach ($results as $migration => $result) {
-            if ($result === false) {
-                $result = 'KO';
-            } elseif ($result === true) {
-                $result = 'OK';
-            }
-
-            if ($result !== 'OK') {
-                $has_error = true;
-            }
-
-            $results_as_text[] = "{$migration}: {$result}";
-        }
-        return Response::text($has_error ? 500 : 200, implode("\n", $results_as_text));
     }
 }
